+++ examples/get_userid.py
@@ -12,27 +12,36 @@
 # See the License for the specific language governing permissions and
 # limitations under the License
 
-"""Fetch the text for a Rel source from a specific database"""
+"""Get the userid corresponding to the given user email."""
 
 from argparse import ArgumentParser
-from railib import api, config
-from show_error import show_error
+import json
+from urllib.request import HTTPError
+from railib import api, config, show
 
 
-@show_error
-def run(database: str, engine: str, source: str, profile: str):
+def get_userid(ctx, email: str) -> str:
+    rsp = api.list_users(ctx)
+    for item in rsp:
+        if item["email"] == email:
+            return item["id"]
+    return None
+
+
+def run(email: str, profile: str):
     cfg = config.read(profile=profile)
     ctx = api.Context(**cfg)
-    rsp = api.get_source(ctx, database, engine, source)
-    print(rsp)
+    rsp = get_userid(ctx, email)
+    print(json.dumps(rsp, indent=2))
 
 
 if __name__ == "__main__":
     p = ArgumentParser()
-    p.add_argument("database", type=str, help="database name")
-    p.add_argument("engine", type=str, help="engine name")
-    p.add_argument("source", type=str, help="source name")
     p.add_argument("-p", "--profile", type=str,
                    help="profile name", default="default")
+    p.add_argument("email", type=str, help="user email")
     args = p.parse_args()
-    run(args.database, args.engine, args.source, args.profile)
+    try:
+        run(args.email, args.profile)
+    except HTTPError as e:
+        show.http_error(e)